split typescript file fix some eslint complaints remove circular dependency between characters and options
