+++ .eslintrc.json
@@ -8,7 +8,7 @@
     "parserOptions": {
         "ecmaVersion": "latest",
         "sourceType": "module",
-        "project": "./tsconfig.json"
+        "project": "tsconfig.json"
     },
     "rules": {
         "quote-props": "off",
